Move domain compilation to dedicated class
