+++ service/src/main/kotlin/org/opensearch/simpleschema/action/SimpleSchemaDomainActions.kt
@@ -4,6 +4,7 @@ import org.opensearch.OpenSearchStatusException
 import org.opensearch.commons.authuser.User
 import org.opensearch.rest.RestStatus
 import org.opensearch.simpleschema.SimpleSchemaPlugin
+import org.opensearch.simpleschema.domain.DomainCompiler
 import org.opensearch.simpleschema.domain.DomainRepository
 import org.opensearch.simpleschema.domain.DomainResource
 import org.opensearch.simpleschema.index.SimpleSearchIndex
@@ -40,7 +41,7 @@ internal object SimpleSchemaDomainActions {
             )
         }
         // TODO unaddressed edge case where compilation is successful but storage fails
-        compile(objectDoc, user)
+        DomainCompiler.compile(objectDoc, user)
         val docId = SimpleSearchIndex.createSimpleSchemaObject(objectDoc, request.objectId)
         docId ?: throw OpenSearchStatusException(
             "Object creation failed",
@@ -60,30 +61,5 @@ internal object SimpleSchemaDomainActions {
         return GetSimpleSchemaDomainResponse(result.simpleSchemaObjectDoc.objectId)
     }
 
-    private fun compile(objectDoc: SimpleSchemaObjectDoc, user: User?) {
-        if (objectDoc.type != SimpleSchemaObjectType.SCHEMA_DOMAIN) {
-            throw IllegalArgumentException("Attempted to domain-compile a non-domain object doc: " +
-                "expected type ${SimpleSchemaObjectType.SCHEMA_DOMAIN} but got ${objectDoc.type}")
-        }
-        val compilationData = objectDoc.objectData as SchemaCompilationType
-        val entityData = getEntityData(compilationData.entities, user)
-        if (entityData.keys != compilationData.entities.toSet()) {
-            throw OpenSearchStatusException(
-                "Compilation failed: One or more required entities could not be found.",
-                RestStatus.NOT_FOUND
-            )
-        }
-        // TODO: generating domain resource,
-        //  but skipping compilation until more details about the expected artifact are known
-        val domain = DomainResource(compilationData.name, compilationData.entities)
-        DomainRepository.createDomain(domain)
-    }
 
-    private fun getEntityData(entities: List<String>, user: User?): Map<String, SimpleSchemaObjectDoc> {
-        return SimpleSchemaActions
-            .get(GetSimpleSchemaObjectRequest(entities.toSet()), user)
-            .searchResult
-            .objectList
-            .associateBy { it.objectId }
-    }
 }
